Fix example local.settings.json
The example local.settings.json was mapped case insensitive compared to the definitions in code. This is acceptable for the inferior operating system Windows, as it operates as loose as the loose units using it (lol). But alas, when we are using more advanced operating systems from the future, the settings never set. This remedies that.