Added "Guests" functionality
@KelvinTegelaar

This is slightly broken
"View" guest does not work
"Edit" guest does not work

I had to add the additional "GuestDetails" and "GuestLastLoginDetails" because the "Userdetails" and "UserLastLoginDetails" both refer back to the "ListUsers" function which no longer contains "Guests"

If we wanted to keep "Guests" showing in the main "User" list, then a lot of what i have done can be undone, and the "View" and "Edit" button are likely to then work, but i dont think we should keep the guests in the "User" list now that we have a separate place for them

Once the "View" and "Edit" pages are working there may well be some clean up required on them which i am happy to do. But its hard to do that whilst its not working.

Sorry, i hope this isnt too much of a mess!